Create Sequence Equation.py
The problem involves finding a sequence of integers y such that for each x  from 1 to n, p[p[y]]=x, where p is a 1-indexed permutation of integers from 1 to n. By determining the positions of x and its intermediate index in p, the solution constructs the sequence y efficiently for all x.